make errors getting issues non-fatal
This can happen if someone moves a repo between the previous merge and then a release. The Issue number is the only thing that is used to get issue details but really the full path should be used (if at all possible). This change just prevents this situation from causing an unrecoverable error